+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- changed location of index.d.ts file
+
+### Added
+
+- Added typing for the next config
+- added new parameter `enabled` to determ if google secrets should be loaded or not
+
 ### Changed
 
 - changed `projects/`-part of the project name to optional